Initial checkin.
Basic map handling works. Draws a boat, and a speed vector for the boat.
'+' and '-' keys zoom the map, 'l' and 'r' rotate the map. Dragging it pans.

Next task is to add Websocket communication between server and web page,
letting the server simulate a moving boat and having the map update.